requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
six==1.16.0
tzdata==2024.1
Werkzeug==3.0.2
xlrd==2.0.1
zipp==3.18.1
